Split waypoint draft management into a nav-info draft management
